Fixed Distributed enemy, Refactored other enemies, added shooter enemy, redid map, tweeked player
Fixed distributed enemy so that all clones die when main one dies. Refactored other enemies to die by the player's weapon rather than jumping on the head. Added a shooter enemy that chases the player and shoots them. Tweeked the player to get the shooting to work better. Redid the map so the focus is more on fighting rather than trying to make the jumps and platforming.